SVfitStandaloneAlgorithm: split code into two files
